GiphyAPI and GiphyCLI with api test only
